+++ example/lib/screens/check_mark_indicator_screen.dart
@@ -4,6 +4,8 @@ import 'package:example/widgets/example_list.dart';
 import 'package:flutter/material.dart';
 
 class CheckMarkIndicatorScreen extends StatefulWidget {
+  const CheckMarkIndicatorScreen({Key? key}) : super(key: key);
+
   @override
   _CheckMarkIndicatorScreenState createState() =>
       _CheckMarkIndicatorScreenState();
@@ -12,12 +14,12 @@ class CheckMarkIndicatorScreen extends StatefulWidget {
 class _CheckMarkIndicatorScreenState extends State<CheckMarkIndicatorScreen> {
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
+    return const Scaffold(
       backgroundColor: appBackgroundColor,
-      appBar: const ExampleAppBar(),
+      appBar: ExampleAppBar(),
       body: SafeArea(
         child: CheckMarkIndicator(
-          child: const ExampleList(),
+          child: ExampleList(),
         ),
       ),
     );