+++ example/lib/screens/example_indicator_screen.dart
@@ -5,6 +5,8 @@ import 'package:example/widgets/example_list.dart';
 import 'package:flutter/material.dart';
 
 class ExampleIndicatorScreen extends StatelessWidget {
+  const ExampleIndicatorScreen({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     final CustomIndicatorConfig customIndicator =