+++ example/lib/widgets/example_app_bar.dart
@@ -3,18 +3,19 @@ import 'package:flutter/material.dart';
 const appBackgroundColor = Color(0xFFf8f4fc);
 
 class ExampleAppBar extends StatelessWidget implements PreferredSizeWidget {
-  const ExampleAppBar();
+  const ExampleAppBar({Key? key}) : super(key: key);
+
   @override
   Widget build(BuildContext context) {
     return AppBar(
-      iconTheme: IconThemeData(
-        color: const Color(0xff877162),
+      iconTheme: const IconThemeData(
+        color: Color(0xff877162),
       ),
       backgroundColor: appBackgroundColor,
-      title: Text(
+      title: const Text(
         "flutter_custom_refresh_indicator",
         style: TextStyle(
-          color: const Color(0xff877162),
+          color: Color(0xff877162),
         ),
       ),
       elevation: 3,
@@ -22,5 +23,5 @@ class ExampleAppBar extends StatelessWidget implements PreferredSizeWidget {
   }
 
   @override
-  Size get preferredSize => Size.fromHeight(kToolbarHeight);
+  Size get preferredSize => const Size.fromHeight(kToolbarHeight);
 }