+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - Added `onStateChanged` function argument that allows tracking indicator state changes.
   - The `IndicatorStateHelper` class is now deprecated in favor of `onStateChange` function and `IndicatorStateChange` class.
   - Initial support for programmatically-controlled indicators has been added. Added the `show`,` hide` and `refresh` methods to the` CustomRefreshIndicatorState` class. It can be accessed via GlobalKey. Take a look at an [programmatically-controlled screen example](/example/lib/screens/programmatically_controlled_indicator_screen.dart).
+  - Use the `flutter_lints` package for analysis.
 
 ## 1.0.0